requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-adminlte3==0.1.6
django-bootstrap4==3.0.1
django-bootstrap5==2.1.1
django-cpf==0.1.0
django-crispy-forms==1.14.0
django-upload-validator==1.1.5
Faker==8.12.0
idna==3.2
mixer==7.1.2
oauthlib==3.1.1
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python-magic==0.4.25
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
